requirements.txt
langchain==0.3.28
langchain-community==0.3.28
langchain-google-genai==2.1.12
langchain-openai==0.2.14
langchain-weaviate==0.0.6
langchain-text-splitters==0.3.11
weaviate-client==4.16.2
python-dotenv==1.2.2
pydantic==2.12.5
pydantic-settings==2.13.1
requests==2.33.1
pandas==3.0.2
numpy==2.4.4
opentelemetry-api==1.40.0
fastapi==0.135.2
uvicorn==0.42.0
python-multipart==0.0.22
